v16-e2e/references/libraries/order_management/requirements.txt
SQLAlchemy==2.0.25
python-dotenv==1.0.0
